+++ .github/workflows/deploy.yml
@@ -49,7 +49,7 @@ jobs:
         run: just analyser::run
 
       - name: Copy generated files to github pages folder
-        run: cp -r analyser/generated_markdown/ github-pages
+        run: cp -r analyser/application/generated_markdown/ github-pages
 
       - name: Build with Jekyll
         uses: actions/jekyll-build-pages@v1
